i#1312 AVX-512 support: Add tuple type to decode table. (#3706)
Adds the tuple type from Intel's Vol.2A 2.6.5 "Compressed Displacement (disp8*N) Support
in EVEX" to the decode tables.

Extends flags field of instr_info_t to 32-bit. The upper half is reserved for the evex
tuple type of AVX-512.

Tests will be added in future patch. While each tuple type has been carefully added to
the decode tables, there is no decoder/encoder support yet and no tests.

Issue: #1312

+++ core/arch/x86/encode.c
@@ -1095,7 +1095,7 @@ opnd_needs_evex(opnd_t opnd)
 
 static bool
 opnd_type_ok(decode_info_t *di /*prefixes field is IN/OUT; x86_mode is IN*/, opnd_t opnd,
-             int optype, opnd_size_t opsize, ushort flags)
+             int optype, opnd_size_t opsize, uint flags)
 {
     DOLOG(ENC_LEVEL, LOG_EMIT, {
         dcontext_t *dcontext = get_thread_private_dcontext();
